src/requirements.txt
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
comtypes==1.1.11
Flask==2.0.2
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.1.0
MarkupSafe==2.0.1
MouseInfo==0.1.3
nltk==3.7
Pillow==9.0.1
playsound==1.2.2
pocketsphinx==0.1.15
PyAudio=../ref/PyAudio-0.2.11-cp310-cp310-win_amd64.whl
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.28
python-decouple==3.6
pyttsx3==2.90
pytweening==1.0.4
pywhatkit==5.3
pywin32==303
regex==2022.1.18
requests==2.27.1
soupsieve==2.3.1
SpeechRecognition==3.8.1
textblob==0.17.1
tqdm==4.62.3
urllib3==1.26.8
Werkzeug==2.0.3
wikipedia==1.4.0
windows-curses==2.3.0
